Pretty massive changes.
I removed the separate map search for iphone. You only get the map for ipad. That's because it really needs to be shown with the simple buttons to be effective. iphone screen is just too small.

This also completely simplifies the storyboards. No more custom segues. They were fun, but won't be missed.

The map and list searches seem to be operating at peak performance.

I moved the storyboards into the variant. I may also move the entire localization file there, and eliminate the Perl script.