OK. I "punted" on the whole details leak thing.
I figgered that it was buried somewhere deep inside the MapKit. I did everything I could on my end to solve any leaks, and had to accept that I wasn't gonna find it.

So I set up the details screen as a reusable view controller.

I am such a wimp...